+++ src/test/java/ortus/boxlang/modules/compat/IntegrationTest.java
@@ -5,6 +5,8 @@ import static com.google.common.truth.Truth.assertThat;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import ortus.boxlang.runtime.scopes.Key;
+
 /**
  * This loads the module and runs an integration test on the module.
  */
@@ -14,5 +16,10 @@ public class IntegrationTest extends BaseIntegrationTest {
 	@Test
 	public void testModuleLoads() {
 		assertThat( moduleService.getRegistry().containsKey( moduleName ) ).isTrue();
+
+		// Verify that the CFIDE mapping was registered
+		assertThat(
+		    runtime.getConfiguration().hasMapping( Key.of( "CFIDE" ) )
+		).isTrue();
 	}
 }